+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0]- 2020-03-24
+
+### Added
+
+- `Select` and `RadioGroup` to override components
+
+### Changed
+
+- `Select` component to also accept number as a value
+- Updated dev dependencies
+
+### Fixed
+
+- `Counter` value handling
+- [PR](https://github.com/jeslage/storybook-addon-theme-playground/pull/1): Bump acorn from 6.4.0 to 6.4.1.
+
 ## [1.2.2]- 2020-01-26
 
 ### Fixed
@@ -62,7 +78,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.2...develop
+[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.0...develop
+[1.3.0]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.2...1.3.0
 [1.2.2]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.1.0...v1.2.0